feat(automine): move automine mode to types file
